requirements.txt
matplotlib~=3.5.2
numpy~=1.22.4
pillow~=8.4.0
mayavi~=4.7.4
scikit-image~=0.19.2
scipy~=1.8.1